+++ WhistleWindLobotomyMod/CardAppearanceBehaviours/appearanceDragonBackground.cs
@@ -1,4 +1,6 @@
 ﻿using DiskCardGame;
+using InscryptionAPI.PixelCard;
+using UnityEngine;
 using WhistleWind.Core.Helpers;
 
 
@@ -13,22 +15,22 @@ namespace WhistleWindLobotomyMod
             DragonBodyBackground.appearance = CardHelper.CreateAppearance<DragonBodyBackground>(pluginGuid, "DragonBodyBackground").Id;
         }
     }
-    public class DragonHeadBackground : CardAppearanceBehaviour
+    public class DragonHeadBackground : PixelAppearanceBehaviour
     {
         public static Appearance appearance;
-        public override void ApplyAppearance() =>
-            base.Card.RenderInfo.baseTextureOverride = TextureLoader.LoadTextureFromFile("dragonHeadBackground");
+        public override Sprite OverrideBackground() => TextureLoader.LoadSpriteFromFile("dragonHeadBackground_pixel");
+        public override void ApplyAppearance() => base.Card.RenderInfo.baseTextureOverride = TextureLoader.LoadTextureFromFile("dragonHeadBackground");
     }
-    public class DragonHornsBackground : CardAppearanceBehaviour
+    public class DragonHornsBackground : PixelAppearanceBehaviour
     {
         public static Appearance appearance;
-        public override void ApplyAppearance() =>
-            base.Card.RenderInfo.baseTextureOverride = TextureLoader.LoadTextureFromFile("dragonHornsBackground");
+        public override Sprite OverrideBackground() => TextureLoader.LoadSpriteFromFile("dragonHornsBackground_pixel");
+        public override void ApplyAppearance() => base.Card.RenderInfo.baseTextureOverride = TextureLoader.LoadTextureFromFile("dragonHornsBackground");
     }
-    public class DragonBodyBackground : CardAppearanceBehaviour
+    public class DragonBodyBackground : PixelAppearanceBehaviour
     {
         public static Appearance appearance;
-        public override void ApplyAppearance() =>
-            base.Card.RenderInfo.baseTextureOverride = TextureLoader.LoadTextureFromFile("dragonBodyBackground");
+        public override Sprite OverrideBackground() => TextureLoader.LoadSpriteFromFile("dragonBodyBackground_pixel");
+        public override void ApplyAppearance() => base.Card.RenderInfo.baseTextureOverride = TextureLoader.LoadTextureFromFile("dragonBodyBackground");
     }
 }