+++ WhistleWindLobotomyMod/CardAppearanceBehaviours/appearanceEventCardBackground.cs
@@ -1,4 +1,6 @@
 ﻿using DiskCardGame;
+using InscryptionAPI.PixelCard;
+using UnityEngine;
 using WhistleWind.Core.Helpers;
 
 
@@ -12,14 +14,16 @@ namespace WhistleWindLobotomyMod
             RareEventBackground.appearance = CardHelper.CreateAppearance<RareEventBackground>(pluginGuid, "EventCardBackgroundRare").Id;
         }
     }
-    public class EventBackground : CardAppearanceBehaviour
+    public class EventBackground : PixelAppearanceBehaviour
     {
         public static Appearance appearance;
+        public override Sprite OverrideBackground() => TextureLoader.LoadSpriteFromFile("eventCardBackground_pixel");
         public override void ApplyAppearance() => base.Card.RenderInfo.baseTextureOverride = TextureLoader.LoadTextureFromFile("eventCardBackground");
     }
-    public class RareEventBackground : CardAppearanceBehaviour
+    public class RareEventBackground : PixelAppearanceBehaviour
     {
         public static Appearance appearance;
+        public override Sprite OverrideBackground() => TextureLoader.LoadSpriteFromFile("eventCardBackgroundRare_pixel");
         public override void ApplyAppearance() => base.Card.RenderInfo.baseTextureOverride = TextureLoader.LoadTextureFromFile("eventCardBackgroundRare");
     }
 }