+++ WhistleWindLobotomyMod/Cards/willBeBadWolf.cs
@@ -18,9 +18,10 @@ namespace WhistleWindLobotomyMod
                 willBeBadWolf,
                 "Big and Will Be Bad Wolf",
                 "It's the fate of all wolves to be the villains of fairy tales.",
-                attack: 1, health: 3, blood: 2)
+                attack: 3, health: 4, blood: 3)
                 .SetPortraits(willBeBadWolf)
                 .AddAbilities(Assimilator.ability)
+                .AddSpecialAbilities(CrimsonScar.specialAbility)
                 .AddTribes(Tribe.Canine);
 
             CreateCard(willBeBadWolfCard, CardHelper.ChoiceType.Common, RiskLevel.Waw);