+++ WhistleWindLobotomyMod/Patches/ExtendedSniperLogicPatches.cs
@@ -6,30 +6,33 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using WhistleWind.Core;
+using WhistleWind.Core.Helpers;
+using WhistleWindLobotomyMod;
+using WhistleWindLobotomyMod.Core;
 using static InscryptionCommunityPatch.Card.SniperFix;
 
 // custom version of the Sniper fix that adds Marksman compatibility and the special behaviour of Blue Star and Judgement Bird
 namespace WhistleWind.AbnormalSigils.Patches
 {
     [HarmonyPatch]
-    internal class AddSniperPiperPatch
+    internal class ExtendedSniperLogicPatches
     {
-        private const string singleton = "wstl:JudgementExecution";
+        private const string executeId = "wstl:JudgementExecution";
 
         [HarmonyPrefix]
         [HarmonyPatch(typeof(CombatPhaseManager), nameof(CombatPhaseManager.VisualizeConfirmSniperAbility))]
         [HarmonyPatch(typeof(Part1SniperVisualizer), nameof(Part1SniperVisualizer.VisualizeConfirmSniperAbility))]
         private static bool PerformHanging(CardSlot targetSlot)
         {
-            if (targetSlot.Card?.TemporaryMods.Exists(x => x.singletonId == singleton) ?? false)
+            if (targetSlot.Card?.TemporaryMods.Exists(x => x.singletonId == executeId) ?? false)
                 targetSlot.Card.Anim.SetMarkedForSacrifice(marked: true);
 
             return true;
         }
         [HarmonyPrefix, HarmonyPatch(typeof(SniperFix), nameof(SniperFix.SniperAttack))]
-        private static bool OverrideWithMarksman(CombatPhaseManager instance, CardSlot slot, ref IEnumerator __result)
+        private static bool OverrideSniperSequence(CombatPhaseManager instance, CardSlot slot, ref IEnumerator __result)
         {
-            if (IsJudgementBird(slot))
+            if (IsJudgementBird(slot) || IsEnraged(slot) || IsRaging(slot))
             {
                 __result = WstlSniperSequence(instance, slot);
                 return false;
@@ -39,48 +42,70 @@ namespace WhistleWind.AbnormalSigils.Patches
         [HarmonyPostfix, HarmonyPatch(typeof(SniperFix), nameof(SniperFix.WillDieFromSharp))]
         private static void AddExtraChecks(CardSlot slot, ref bool __result)
         {
-            if (IsJudgementBird(slot)) // Judgement Bird does not trigger OnDealDamage or OnTakeDamage
-            {
+            // Judgement Bird does not trigger OnDealDamage or OnTakeDamage
+            if (IsJudgementBird(slot))
                 __result = false;
-                return;
-            }
-
-            if (slot.Card.HasAbility(Punisher.ability))
-                __result = true;
         }
         private static IEnumerator WstlSniperSequence(CombatPhaseManager instance, CardSlot slot)
         {
-            List<CardSlot> opposingSlots = slot.Card.GetOpposingSlots();
+            List<CardSlot> opposingSlots = new(); // the slots to attacks
             Singleton<ViewManager>.Instance.SwitchToView(Singleton<BoardManager>.Instance.CombatView, false, false);
             Singleton<ViewManager>.Instance.Controller.LockState = ViewLockState.Locked;
             int numAttacks = GetAttackCount(slot.Card);
-
-            opposingSlots.Clear();
             Singleton<ViewManager>.Instance.Controller.SwitchToControlMode(Singleton<BoardManager>.Instance.ChoosingSlotViewMode, false);
             Singleton<ViewManager>.Instance.Controller.LockState = ViewLockState.Unlocked;
-
             Part1SniperVisualizer visualizer = null;
             if (SaveManager.SaveFile.IsPart1)
                 visualizer = instance.GetComponent<Part1SniperVisualizer>() ?? instance.gameObject.AddComponent<Part1SniperVisualizer>();
 
-            if (slot.Card.OpponentCard)
-                yield return WstlOpponentLogic(instance, visualizer, opposingSlots, slot, numAttacks);
+            if (IsRaging(slot))
+            {
+                opposingSlots = slot.Card.GetOpposingSlots();
+                for (int i = 0; i < opposingSlots.Count; i++)
+                {
+                    instance.VisualizeConfirmSniperAbility(opposingSlots[i]);
+                    visualizer?.VisualizeConfirmSniperAbility(opposingSlots[i]);
+                }
+            }
             else
-                yield return WstlPlayerLogic(instance, visualizer, opposingSlots, slot, numAttacks);
+            {
+                CardSlot enragedTarget = null;
+                if (IsEnraged(slot))
+                {
+                    string nameToFind;
+                    if (slot.Card.Info.name == "wstl_redHoodedMercenary")
+                        nameToFind = "wstl_willBeBadWolf";
+                    else
+                        nameToFind = "wstl_redHoodedMercenary";
+
+                    enragedTarget = BoardManager.Instance.AllSlotsCopy.Find(x => x.Card?.Info.name == nameToFind);
+                    if (enragedTarget != null)
+                    {
+                        for (int i = 0; i < numAttacks; i++)
+                            opposingSlots.Add(enragedTarget);
+
+                        instance.VisualizeConfirmSniperAbility(enragedTarget);
+                        visualizer?.VisualizeConfirmSniperAbility(enragedTarget);
+                    }
+                }
+
+                if (enragedTarget == null)
+                {
+                    if (slot.Card.OpponentCard)
+                        yield return WstlOpponentLogic(instance, visualizer, opposingSlots, slot, numAttacks);
+                    else
+                        yield return WstlPlayerLogic(instance, visualizer, opposingSlots, slot, numAttacks);
+                }
+            }
 
             Singleton<ViewManager>.Instance.Controller.SwitchToControlMode(Singleton<BoardManager>.Instance.DefaultViewMode, false);
             Singleton<ViewManager>.Instance.Controller.LockState = ViewLockState.Locked;
             foreach (CardSlot opposingSlot in opposingSlots)
             {
                 Singleton<ViewManager>.Instance.SwitchToView(Singleton<BoardManager>.Instance.CombatView, false, false);
-                if (opposingSlot.Card != null && opposingSlot.Card.LacksAllTraits(Trait.Terrain, Trait.Pelt, AbnormalPlugin.ImmuneToInstaDeath) && !opposingSlot.Card.AttackIsBlocked(slot))
+                if (IsJudgementBird(slot) && opposingSlot.Card != null && !ImmuneToHanging(opposingSlot) && !opposingSlot.Card.AttackIsBlocked(slot))
                 {
-                    if (slot.Card.FaceDown)
-                    {
-                        slot.Card.SetFaceDown(false);
-                        slot.Card.UpdateFaceUpOnBoardEffects();
-                        yield return new WaitForSeconds(0.2f);
-                    }
+                    slot.Card.FlipFaceDown(false);
                     slot.Card.Anim.StrongNegationEffect();
                     yield return new WaitForSeconds(0.4f);
                     yield return Execution(opposingSlot.Card);
@@ -111,10 +136,9 @@ namespace WhistleWind.AbnormalSigils.Patches
                     delegate (CardSlot s)
                     {
                         opposingSlots.Add(s);
-                        if (!ImmuneToHanging(s))
-                        {
-                            s.Card.AddTemporaryMod(new() { singletonId = singleton });
-                        }
+                        if (IsJudgementBird(slot) && !ImmuneToHanging(s))
+                            s.Card.AddTemporaryMod(new() { singletonId = executeId });
+
                         instance.VisualizeConfirmSniperAbility(s);
                         visualizer?.VisualizeConfirmSniperAbility(s);
                     }, null, delegate (CardSlot s)
@@ -152,8 +176,8 @@ namespace WhistleWind.AbnormalSigils.Patches
                 }
 
                 opposingSlots.Add(attackSlot);
-                if (!ImmuneToHanging(attackSlot))
-                    attackSlot.Card.AddTemporaryMod(new() { singletonId = "wstl:JudgementExecution" });
+                if (IsJudgementBird(slot) && !ImmuneToHanging(attackSlot))
+                    attackSlot.Card.AddTemporaryMod(new() { singletonId = executeId });
 
                 instance.VisualizeConfirmSniperAbility(attackSlot);
                 visualizer?.VisualizeConfirmSniperAbility(attackSlot);
@@ -161,7 +185,9 @@ namespace WhistleWind.AbnormalSigils.Patches
             }
         }
         private static bool ImmuneToHanging(CardSlot slot) => slot.Card?.HasAnyOfTraits(Trait.Terrain, Trait.Pelt, AbnormalPlugin.ImmuneToInstaDeath) ?? true;
-        private static bool IsJudgementBird(CardSlot slot) => slot.Card.HasTrait(AbnormalPlugin.Executioner);
+        private static bool IsJudgementBird(CardSlot slot) => slot.Card.HasTrait(LobotomyCardManager.TraitExecutioner);
+        private static bool IsEnraged(CardSlot slot) => slot.Card.GetComponent<CrimsonScar>()?.Enraged ?? false;
+        private static bool IsRaging(CardSlot slot) => slot.Card.HasSpecialAbility(BlindRage.specialAbility);
         private static IEnumerator Execution(PlayableCard target)
         {
             target.Anim.PlaySacrificeSound();